+++ python/rpmmodule.c
@@ -540,5 +540,5 @@ void init_rpm(void)
 
     REGISTER_ENUM(RPMDBI_PACKAGES);
 
-    REGISTER_ENUM((int)RPMAL_NOMATCH);
+    REGISTER_ENUM((long)RPMAL_NOMATCH);
 }
